CloudFunctions/GCPSQLAccessSample/requirements.txt
pandas
PyMySQL
sqlalchemy
google-cloud-logging